+++ frontend/src/components/Listings/ListingsCard.jsx
